cdp_avp: Rename module interface file #82
