+++ src/modules/ims_charging/ccr.c
@@ -1,4 +1,4 @@
-#include "../cdp_avp/mod_export.h"
+#include "../cdp_avp/cdp_avp_mod.h"
 
 #include "ccr.h"
 #include "Ro_data.h"
