+++ src/modules/ims_charging/diameter_ro.h
@@ -2,7 +2,7 @@
 #define __CLIENT_RF_DIAMETER_RO_H
 
 #include "../cdp/cdp_load.h"
-#include "../cdp_avp/mod_export.h"
+#include "../cdp_avp/cdp_avp_mod.h"
 
 int AAASendCCR(AAASession * session);
