+++ src/modules/ims_ocs/mod.c
@@ -22,7 +22,7 @@
 #include "../../core/sr_module.h"
 #include "../../core/route.h"
 #include "../cdp/cdp_load.h"
-#include "../cdp_avp/mod_export.h"
+#include "../cdp_avp/cdp_avp_mod.h"
 #include "../../core/parser/msg_parser.h"
 #include "mod.h"
 #include "msg_faker.h"
