+++ src/modules/ims_ocs/msg_faker.c
@@ -23,7 +23,7 @@
 #include "../../core/parser/msg_parser.h"
 #include "../../core/globals.h"
 #include "../cdp/cdp_load.h"
-#include "../cdp_avp/mod_export.h"
+#include "../cdp_avp/cdp_avp_mod.h"
 #include "ocs_avp_helper.h"
 
 #include "msg_faker.h"
