+++ src/modules/ims_qos/rx_avp.c
@@ -49,7 +49,7 @@
 
 
 #include <arpa/inet.h>
-#include "../cdp_avp/mod_export.h"
+#include "../cdp_avp/cdp_avp_mod.h"
 #include "../../modules/ims_dialog/dlg_load.h"
 #include "../ims_usrloc_pcscf/usrloc.h"
 #include "rx_authdata.h"
