+++ src/modules/ims_qos/rx_str.c
@@ -50,7 +50,7 @@
 
 
 
-#include "../cdp_avp/mod_export.h"
+#include "../cdp_avp/cdp_avp_mod.h"
 #include "../../modules/ims_dialog/dlg_load.h"
 #include "../ims_usrloc_pcscf/usrloc.h"
 #include "rx_authdata.h"
